Перенес функции для создания данных в main.js
